services/api/pyproject.toml
[project]
name = "neuroforge-api"
version = "0.1.0"
description = "Patient-anchored medical research platform for neuronal regrowth / NGF investigations"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "pypdf>=4.0",
    "feedparser>=6.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
ingest = [
    "python-docx>=1.1",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1",
    "striprtf>=0.0.26",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
